\name{pnmtrem}
\alias{pnmtrem}
\alias{pnmtrem}
\docType{package}
\title{
Probit-Normal Marginalized Transition Random Effects Models 
}
\description{
Fits Probit-Normal Marginalized Transition Random Effects Models which is proposed for modeling multivariate longitudinal binary data by

Asar, O., Ilk, O., Sezer, A. D. (2013). Probit Normal Marginalized Transition Random Effects Models. Tentatively accepted.
}
\details{
\tabular{ll}{
Package: \tab pnmtrem\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2013-02-03\cr
License: \tab GPL (>=2)\cr
}
%\tabular{ll}{
%Package: \tab pnmtrem\cr
%Type: \tab Package\cr
%Version: \tab 1.2\cr
%Date: \tab 2012-02-03\cr
%License: \tab GPL (>=2)\cr
%}
}
