\name{pnmtrem}
\alias{pnmtrem}
\alias{pnmtrem}
\docType{package}
\title{
Probit-Normal Marginalized Transition Random Effects Models 
}
\description{
Fits Probit-Normal Marginalized Transition Random Effects Models which is proposed for modeling multivariate longitudinal binary data by

Asar, O., Ilk, O., Sezer, A. D. (2013). A marginalized multilevel model for analyzing multivariate longitudinal binary data. Submitted.
}
\details{
\tabular{ll}{
Package: \tab pnmtrem\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2013-05-19\cr
License: \tab GPL (>=2)\cr
}
}
