\name{pnmtrem}
\alias{pnmtrem}
\alias{pnmtrem}
\docType{package}
\title{
Probit-Normal Marginalized Transition Random Effects Models 
}
\description{
Fits Probit-Normal Marginalized Transition Random Effects Models which is proposed for modeling multivariate longitudinal binary data by

Asar, O., Ilk, O., Sezer, A. D. (2012). Probit Normal Marginalized Transition Random Effects Models. Submitted.
}
\details{
\tabular{ll}{
Package: \tab pnmtrem\cr
Type: \tab Package\cr
%Version: \tab 1.1\cr
%Date: \tab 2012-09-10\cr
%License: \tab GPL (>=2)\cr
}
\tabular{ll}{
%Package: \tab pnmtrem\cr
%Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-09-10\cr
License: \tab GPL (>=2)\cr
}
}
